change bash global var to use class
